+++ src/clubRules.ts
@@ -1,6 +1,7 @@
 export const defaultClubRules = [
-    'Nur Vereinsmitglieder dürfen Plätze reservieren und auf ihnen spielen. Gäste dürfen gemeinsam mit einem Vereinsmitglied für 10 Euro pro Stunde spielen, höchstens jedoch dreimal pro Jahr.',
+    'Nur Vereinsmitglieder dürfen die Tennisplätze reservieren und darauf spielen. Gäste können in Begleitung eines Vereinsmitglieds gegen eine Gebühr und in begrenztem Umfang ebenfalls spielen.',
     'Erscheint innerhalb der ersten 10 Minuten nach Beginn der Reservierungszeit kein Spieler auf dem Platz, gilt die Reservierung als storniert. In diesem Fall kann jeder Spieler den Platz für die verbleibende Zeit bis zur nächsten Reservierung nutzen.',
-    'Spieler sollten ihr Spiel mindestens 5 Minuten vor Ablauf der Reservierungszeit beenden, damit sie genügend Zeit haben, den Platz für die nachfolgenden Spieler vorzubereiten.',
-    'Vor dem Betreten des Platzes müssen die Spieler Sandplatzschuhe anziehen. Cross-Trainer, Laufschuhe und Schuhe mit grobstolligem Profil sind strengstens verboten, da sie den Platzbelag beschädigen.',
+    'Die Spieler sollten ihr Spiel mindestens 5 Minuten vor Ablauf der Reservierungszeit beenden, damit sie genügend Zeit haben, den Platz zu wässern und abzuziehen.',
+    'Bitte betretet die Tennisplätze ausschließlich mit Tennisschuhen!',
+    'Wenn der Platz trocken ist, sollten Sie ihn gründlich wässern. Je nach Temperatur wässern Sie den Platz bitte vor Spielbeginn, nach einem Match und – falls erforderlich – in der Satzpause.',
 ];